third_party/blink/web_tests/external/wpt/permissions-policy/payment-extension-allowed-by-permissions-policy-attribute.https.sub-expected.txt

This is a testharness.js-based test.
FAIL permissions policy "payment" is not supported in cross-origin iframe without allow="payment" attribute assert_equals: NotSupportedError#The 'payment' feature is not enabled in this document. Permissions Policy may be used to delegate Web Payment capabilities to cross-origin child frames.. expected (string) "NotSupportedError#The 'payment' feature is not enabled in this document. Permissions Policy may be used to delegate Web Payment capabilities to cross-origin child frames." but got (undefined) undefined
FAIL permissions policy "payment" can be enabled in cross-origin iframe using allow="payment" attribute assert_equals: NotSupportedError#The user agent does not support public key credentials.. expected (string) "OK" but got (undefined) undefined
Harness: the test ran to completion.
